move aiter moe to dedicated file
